Correct 6-size.c v2

+++ 0x00-hello_world/6-size.c
@@ -14,11 +14,11 @@ int main(void)
 	long long longLong;
 	float floatNum;
 
-	printf("Size of a char: %zu byte(s)\n", sizeof(character));
-	printf("Size of a int: %zu byte(s)\n", sizeof(integer));
-	printf("Size of a long int: %zu byte(s)\n", sizeof(longInt));
-	printf("Size of a long long: %zu byte(s)\n", sizeof(longLong));
-	printf("Size of a float: %zu byte(s)\n", sizeof(floatNum));
+	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(character));
+	printf("Size of a int: %lu byte(s)\n", (unsigned long)sizeof(integer));
+	printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(longInt));
+	printf("Size of a long long: %lu byte(s)\n", (unsigned long)sizeof(longLong));
+	printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(floatNum));
 
 	return (0);
 }
